requirements.txt
gitpython==3.1.18
requests==2.26.0
langchain
tiktoken 
matplotlib 
seaborn 
tqdm
halo
openai
numpy
pandas